djangoRun/requirements.txt
tensorflow-gpu==1.15.0
nltk
python-twitter
emoji
bs4
requests
